fix: component padding correction
